requirements.txt
h5py==3.7.0
lightning==2.0.0
matplotlib==3.5.1
numpy==1.22.3
omegaconf==2.3.0
pandas==1.4.3
pytorch_lightning==2.0.0
PyWavelets==1.4.1
PyYAML==6.0
rich==13.3.3
scikit_learn==1.2.2
scipy==1.7.3
tensorflow==2.12.0
tensorflow_intel==2.11.0
torch==2.0.0
torchinfo==1.7.2
torchmetrics==0.11.4
torchvision==0.15.1
wandb==0.14.0
xarray==2022.12.0
